fix: Fixed maven dependency issues - fallback on older 4.12.0 okhttp version

+++ backend/src/main/java/dev/fascodes/carRental/storage/service/ImageService.java
@@ -115,7 +115,7 @@ public class ImageService {
                     .stream(
                             file.getInputStream(),
                             file.getSize(),
-                            -1
+                            (long) -1
                     )
                     .contentType(file.getContentType())
                     .build()
